Migrate ResourceCompilerTest to gtest
