+++ framework/test/gtest/CMakeLists.txt
@@ -47,6 +47,7 @@ set(_gtest_tests
   StaticBundleResourceTest.cpp
   FrameworkEventTest.cpp
   TestCounterLatch.cpp
+  ResourceCompilerTest
 )
 
 if(BUILD_SHARED_LIBS)
@@ -66,12 +67,17 @@ set(_additional_srcs
 # Build the main test driver executable
 #-----------------------------------------------------------------------------
 
+set(_third_party_srcs 
+  ../../../third_party/miniz.c
+  ../../../third_party/jsoncpp.cpp
+)
+
 # Generate a custom "bundle init" file for the test driver executable
 usFunctionGenerateBundleInit(TARGET ${us_gtest_test_exe_name} OUT _additional_srcs)
 
 usFunctionGetResourceSource(TARGET ${us_gtest_test_exe_name} OUT _additional_srcs)
 
-add_executable(${us_gtest_test_exe_name} ${_gtest_tests} ${_additional_srcs})
+add_executable(${us_gtest_test_exe_name} ${_gtest_tests} ${_additional_srcs} ${_third_party_srcs})
 
 # Needed for ResourceCompilerTest
 add_dependencies(${us_gtest_test_exe_name} ${US_RCC_EXECUTABLE_TARGET})
